Move dids crate into web5 crate
