--- crates/dids/Cargo.toml
@@ -1,29 +1,0 @@
-[package]
-name = "dids"
-version = "0.1.0"
-edition = "2021"
-homepage.workspace = true
-repository.workspace = true
-license-file.workspace = true
-
-[dependencies]
-base64 = { workspace = true }
-byteorder = "1.5.0"
-crypto = { path = "../crypto" }
-did-jwk = "0.1.1"
-did-web = "0.2.2"
-jwk = { path = "../jwk" }
-keys = { path = "../keys" }
-regex = "1.10.4"
-serde = { workspace = true }
-serde_json = { workspace = true }
-simple-dns = "0.7.0"
-ssi-core = "0.1.0"
-ssi-dids = "0.1.1"
-ssi-jwk = "0.1.2"
-thiserror = { workspace = true }
-url = "2.5.0"
-
-[dev-dependencies]
-chrono = { workspace = true }
-tokio = { version = "1.34.0", features = ["macros", "test-util"] }
