+++ crates/web5/Cargo.toml
@@ -8,11 +8,24 @@ license-file.workspace = true
 
 [dependencies]
 base64 = { workspace = true }
+byteorder = "1.5.0"
+did-jwk = "0.1.1"
+did-web = "0.2.2"
 ed25519-dalek = { version = "2.1.1", features = ["rand_core"] }
 getrandom = { version = "0.2.12", features = ["js"] }
 k256 = { version = "0.13.3", features = ["ecdsa", "jwk"] }
 rand = { workspace = true }
+regex = "1.10.4"
 serde = { workspace = true }
 serde_json = { workspace = true }
 sha2 = "0.10.8"
+simple-dns = "0.7.0"
+ssi-core = "0.1.0"
+ssi-dids = "0.1.1"
+ssi-jwk = "0.1.2"
 thiserror = { workspace = true }
+url = "2.5.0"
+
+[dev-dependencies]
+chrono = { workspace = true }
+tokio = { version = "1.34.0", features = ["macros", "test-util"] }
