+++ crates/web5/src/dids/methods/dht/document_packet/rdata_encoder.rs
