+++ crates/web5/src/dids/methods/spruce_mappers/resolution_result.rs
@@ -1,5 +1,7 @@
-use crate::document::Document;
-use crate::resolver::{DocumentMetadata, ResolutionError, ResolutionMetadata, ResolutionResult};
+use crate::dids::document::Document;
+use crate::dids::resolver::{
+    DocumentMetadata, ResolutionError, ResolutionMetadata, ResolutionResult,
+};
 use ssi_dids::did_resolve::{
     DocumentMetadata as SpruceDocumentMetadata, ResolutionMetadata as SpruceResolutionMetadata,
 };
