+++ crates/web5/src/lib.rs
@@ -1,3 +1,4 @@
 pub mod crypto;
+pub mod dids;
 pub mod jwk;
 pub mod keys;
